Move output to Echo class
